+++ src/api/app/index.ts
@@ -1,4 +1,2 @@
 export * from "./types.js";
 export * from "./websocket.js";
-
-// export * from "@holochain/serialization/holochain_serialization_js.js";