+++ src/api/app/websocket.ts
@@ -1,25 +1,9 @@
-/**
- * Defines AppWebsocket, an easy-to-use websocket implementation of the
- * Conductor API for apps
- *
- *    const client = AppWebsocket.connect('ws://127.0.0.1:9000');
- *
- *    client.callZome({...})
- *      .then(() => {
- *        console.log('DNA successfully installed')
- *      })
- *      .catch(err => {
- *        console.error('problem installing DNA:', err)
- *      });
- */
-import { hashZomeCall } from "@holochain/serialization";
-import { decode, encode } from "@msgpack/msgpack";
-import { invoke } from "@tauri-apps/api/tauri";
+import { decode } from "@msgpack/msgpack";
 import Emittery from "emittery";
-import nacl from "tweetnacl";
 import {
   getLauncherEnvironment,
   isLauncher,
+  signZomeCallTauri,
 } from "../../environments/launcher.js";
 import { CapSecret } from "../../hdk/capabilities.js";
 import { InstalledAppId } from "../../types.js";
@@ -32,7 +16,7 @@ import {
   requesterTransformer,
   Transformer,
 } from "../common.js";
-import { getSigningPropsForCell } from "../zome-call-signing.js";
+import { getSigningCredentials, signZomeCall } from "../zome-call-signing.js";
 import {
   AppApi,
   AppInfoRequest,
@@ -50,7 +34,6 @@ import {
   NetworkInfoRequest,
   NetworkInfoResponse,
 } from "./types.js";
-import { getNonceExpiration, randomNonce } from "./util.js";
 
 export class AppWebsocket extends Emittery implements AppApi {
   client: WsClient;
@@ -139,17 +122,6 @@ export class AppWebsocket extends Emittery implements AppApi {
     this._requester("network_info");
 }
 
-interface CallZomeRequestSignedTauri // Tauri requires a number array instead of a Uint8Array
-  extends Omit<
-    CallZomeRequestSigned,
-    "cap_secret" | "cell_id" | "provenance" | "nonce"
-  > {
-  cell_id: [TauriByteArray, TauriByteArray];
-  provenance: TauriByteArray;
-  nonce: TauriByteArray;
-  expires_at: number;
-}
-
 export type Nonce256Bit = Uint8Array;
 
 export interface CallZomeRequestUnsigned extends CallZomeRequest {
@@ -160,18 +132,6 @@ export interface CallZomeRequestUnsigned extends CallZomeRequest {
 
 export interface CallZomeRequestSigned extends CallZomeRequestUnsigned {
   signature: Uint8Array;
-}
-
-type TauriByteArray = number[]; // Tauri requires a number array instead of a Uint8Array
-interface CallZomeRequestUnsignedTauri
-  extends Omit<
-    CallZomeRequestUnsigned,
-    "cap_secret" | "cell_id" | "provenance" | "nonce"
-  > {
-  cell_id: [TauriByteArray, TauriByteArray];
-  provenance: TauriByteArray;
-  nonce: TauriByteArray;
-  expires_at: number;
 }
 
 const callZomeTransform: Transformer<
@@ -187,61 +147,19 @@ const callZomeTransform: Transformer<
       return req;
     }
     if (isLauncher) {
-      const zomeCallUnsigned: CallZomeRequestUnsignedTauri = {
-        provenance: Array.from(req.provenance),
-        cell_id: [Array.from(req.cell_id[0]), Array.from(req.cell_id[1])],
-        zome_name: req.zome_name,
-        fn_name: req.fn_name,
-        payload: Array.from(encode(req.payload)),
-        nonce: Array.from(randomNonce()),
-        expires_at: getNonceExpiration(),
-      };
-
-      const signedZomeCallTauri: CallZomeRequestSignedTauri = await invoke(
-        "sign_zome_call",
-        { zomeCallUnsigned }
-      );
-
-      const signedZomeCall: CallZomeRequestSigned = {
-        provenance: Uint8Array.from(signedZomeCallTauri.provenance),
-        cap_secret: null,
-        cell_id: [
-          Uint8Array.from(signedZomeCallTauri.cell_id[0]),
-          Uint8Array.from(signedZomeCallTauri.cell_id[1]),
-        ],
-        zome_name: signedZomeCallTauri.zome_name,
-        fn_name: signedZomeCallTauri.fn_name,
-        payload: Uint8Array.from(signedZomeCallTauri.payload),
-        signature: Uint8Array.from(signedZomeCallTauri.signature),
-        expires_at: signedZomeCallTauri.expires_at,
-        nonce: Uint8Array.from(signedZomeCallTauri.nonce),
-      };
-
+      const signedZomeCall = await signZomeCallTauri(req);
       return signedZomeCall;
     } else {
-      const signingPropsForCell = getSigningPropsForCell(req.cell_id);
-      if (!signingPropsForCell) {
+      const signingCredentials = getSigningCredentials(req.cell_id);
+      if (!signingCredentials) {
         throw new Error(
-          "cannot sign zome call: signing properties have not been set"
+          "cannot sign zome call: no signing credentials have been authorized"
         );
       }
-      const unsignedZomeCall: CallZomeRequestUnsigned = {
-        ...req,
-        cap_secret: signingPropsForCell.capSecret,
-        provenance: signingPropsForCell.signingKey,
-        payload: encode(req.payload),
-        nonce: randomNonce(),
-        expires_at: getNonceExpiration(),
-      };
-      const hashedZomeCall = await hashZomeCall(unsignedZomeCall);
-      const signature = nacl
-        .sign(hashedZomeCall, signingPropsForCell.keyPair.secretKey)
-        .subarray(0, nacl.sign.signatureLength);
-
-      const signedZomeCall: CallZomeRequestSigned = {
-        ...unsignedZomeCall,
-        signature,
-      };
+      const signedZomeCall = await signZomeCall(
+        signingCredentials,
+        req.payload
+      );
       return signedZomeCall;
     }
   },