+++ src/api/index.ts
@@ -2,4 +2,4 @@ export * from "./admin/index.js";
 export * from "./app-agent/index.js";
 export * from "./app/index.js";
 export { CloneId } from "./common.js";
-export { authorizeNewSigningKeyPair } from "./zome-call-signing.js";
+export * from "./zome-call-signing.js";